pyproject.toml
[project]
name = "fastapi_mongo"
version = "0.1.0"
description = "FastAPI-MongoDB Playground Application"
requires-python = ">=3.12"
dependencies = [
    "fastapi", # Core web framework
    "uvicorn", # ASGI server
    "pandas", # Data processing
    "numpy", # Numerical computations
    "requests", # HTTP client
    "pymongo", # MongoDB driver
    "passlib", # Password hashing
    "python-jose[cryptography]==3.4.0", # JWT handling
    "python-dotenv", # Environment variables
    "email-validator>=2.2.0",
    "pyjwt>=2.10.1",
    "redshift-connector", # AWS Redhift connector,
    "apscheduler", # for periodic background task execution,
    "logging", # logging module,
    "python-json-logger", # json logging,
    "contextvars", # variable storage for async tasks,
    "minio", # minio service enabler
    "ray>=2.43.0",
    "python-multipart>=0.0.20", #for file upload
    "pytest>=8.3.4,<8.5", # for unit testing
    "pytest-asyncio>=0.25.3,<0.26", # for async testing
    "pytest-order==1.3.0",
    "httpx>=0.28.1,<0.29", # for async http requests
    "motor>=3.7.0", # for async mongodb
    "beanie==1.20.0", # for async ODM with mongodb
    "Faker==37.0.1", # Generate fake data for db models
]

[dependency-groups]
dev = [
    "pylint>=3.3.4",
    "pytest>=8.3.4",
]
